Debugged + fixed weird theme coloring issues
Turns out the problem is that the implementation `color-lighten-name` depends on the selected frame... which was not working correctly when initialized by an Emacs server that had not spawned a frame. (I assume it defaulted to TTY colors instead of X colors or something.)

Solution: add a hook that enables the theme the first time that a frame is created by a server.