Change smerge setup
For some reason, my previous config using `M-C` as the smerge prefix key stopped working. `M-C` would get translated to `M-c` automatically.

Switching to `C-.` seems to have solved the problem, since it isn't a keybinding I use in programming modes. Seems like it is bound by flyspell, but I don't use it and I can change that in the future if it becomes a problem.